pyproject.toml
[project]
name = "docling-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.55.1",
    "google-generativeai>=0.8.5",
    "huggingface-hub[hf-xet]>=0.35.3",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
]
